native: example - renaming
